chore: Modify file directory for common card components
