+++ src/components/Card/AddressHashCard.tsx
@@ -1,6 +1,13 @@
+import React, { useContext } from 'react'
+import CopyIcon from '../../assets/copy.png'
+import AppContext from '../../contexts/App'
+import i18n from '../../utils/i18n'
+import { isMobile } from '../../utils/screen'
+import { startEndEllipsis } from '../../utils/string'
+import { copyElementValue } from '../../utils/util'
 import styled from 'styled-components'
 
-export default styled.div`
+const AddressHashCardPanel = styled.div`
   width: 100%;
   border-radius: 6px;
   box-shadow: 2px 2px 6px 0 #dfdfdf;
@@ -66,3 +73,25 @@ export default styled.div`
     }
   }
 `
+
+export default ({ title, hash }: { title: string; hash: string }) => {
+  const appContext = useContext(AppContext)
+  return (
+    <AddressHashCardPanel>
+      <div className="address_hash__title">{title}</div>
+      <div id="address_hash__hash">{isMobile() ? startEndEllipsis(hash, 10) : hash}</div>
+      <div
+        className="address_hash__copy_iocn"
+        role="button"
+        tabIndex={-1}
+        onKeyDown={() => {}}
+        onClick={() => {
+          copyElementValue(document.getElementById('address_hash__hash'))
+          appContext.toastMessage(i18n.t('common.copied'), 3000)
+        }}
+      >
+        <img src={CopyIcon} alt="copy" />
+      </div>
+    </AddressHashCardPanel>
+  )
+}
