--- src/components/Card/TitleCard/index.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import TitleCardPanel from './styled'
-
-export default ({ title }: { title: string }) => {
-  return <TitleCardPanel>{title}</TitleCardPanel>
-}
